init commit

+++ Lab-11/KenaKata/src/main/java/kenakata/order/Chargeable.java
@@ -1,7 +1,7 @@
 package kenakata.order;
 
     public interface Chargeable {
-        long unitCharge();
+        double unitCharge();
 
         long unitVat();
